- update tests - removing Config from clouddirectory to remove circular dependencies - Moved reusable test code that uses fusillade libraries to test.common - Moved reusable test code that does not use fusillade libraries to test.__init__ - Removed unused code from test_api - Using test service accounts in test_api
