+++ tests/common.py
@@ -1,28 +1,18 @@
-import functools
 import json
-import random
-import string
+import os
 import time
-import logging
+
+import jwt
+from dcplib.aws import clients
+from tests import schema_name, random_hex_string
+from fusillade.config import Config
+from fusillade.clouddirectory import publish_schema, create_directory, CloudDirectory
 
 
-from fusillade.clouddirectory import publish_schema, create_directory, cleanup_directory, CloudDirectory
-
-random.seed(time.time())
-
-logger = logging.getLogger()
-schema_name = 'authz'
-
-
-def random_hex_string(length=8):
-    return ''.join([random.choice(string.hexdigits) for i in range(length)])
-
-
-def new_test_directory(directory_name=None) -> CloudDirectory:
-    directory_name = directory_name if directory_name else "test_dir_" + random_hex_string()
-    schema_arn = publish_schema(schema_name, 'T' + random_hex_string())
-    directory = create_directory(directory_name, schema_arn)
-    return directory, schema_arn
+sm = clients.secretsmanager
+service_accounts = json.loads(
+    sm.get_secret_value(SecretId=f"{os.environ['FUS_SECRETS_STORE']}/test_service_accounts")["SecretString"]
+)
 
 
 def create_test_statement(name: str):
@@ -44,28 +34,33 @@ def create_test_statement(name: str):
     return json.dumps(statement)
 
 
-def eventually(timeout: float, interval: float, errors: set = {AssertionError}):
-    """
-    @eventually runs a test until all assertions are satisfied or a timeout is reached.
-    :param timeout: time until the test fails
-    :param interval: time between attempts of the test
-    :param errors: the exceptions to catch and retry on
-    :return: the result of the function or a raised assertion error
-    """
-    def decorate(func):
-        @functools.wraps(func)
-        def call(*args, **kwargs):
-            timeout_time = time.time() + timeout
-            error_tuple = tuple(errors)
-            while True:
-                try:
-                    return func(*args, **kwargs)
-                except error_tuple as e:
-                    if time.time() >= timeout_time:
-                        raise
-                    logger.debug("Error in %s: %s. Retrying after %s s...", func, e, interval)
-                    time.sleep(interval)
+def new_test_directory(directory_name=None) -> CloudDirectory:
+    directory_name = directory_name if directory_name else "test_dir_" + random_hex_string()
+    schema_arn = publish_schema(schema_name, 'T' + random_hex_string())
+    directory = create_directory(directory_name, schema_arn, [service_accounts['admin']['client_email']])
+    return directory
 
-        return call
 
-    return decorate
+def get_service_jwt(service_credentials, email=True, audience=None):
+    iat = time.time()
+    exp = iat + 3600
+    payload = {'iss': service_credentials["client_email"],
+               'sub': service_credentials["client_email"],
+               'aud': audience or Config.audience,
+               'iat': iat,
+               'exp': exp,
+               'scope': ['email', 'openid', 'offline_access']
+               }
+    if email:
+        payload['email'] = service_credentials["client_email"]
+    additional_headers = {'kid': service_credentials["private_key_id"]}
+    signed_jwt = jwt.encode(payload, service_credentials["private_key"], headers=additional_headers,
+                            algorithm='RS256').decode()
+    return signed_jwt
+
+
+def get_auth_header(service_credentials: dict, email=True):
+    info = service_credentials
+    token = get_service_jwt(info, email=email)
+    return {"Authorization": f"Bearer {token}"}
+
